add movie page, add Favourite buttons component
