+++ src/pages/MoviePage/MoviePage.tsx
@@ -3,11 +3,15 @@ import s from "./moviePage.module.css";
 import { useParams } from "react-router-dom";
 import { movieApi } from "../../store/rtkQuery/movieApi";
 import Preloader from "../../ui/elements/Preloader/Preloader";
+import { FavouriteButton } from "../../ui/components/FavouriteButton/FavouriteButton";
 
 export const MoviePage = () => {
   const { id } = useParams();
-  const { data: movieInfo } = movieApi.useGetMovieQuery(id ? parseInt(id) : 0);
-  console.log(movieInfo);
+
+  if (!id) {
+    return <div>Упс, что то пошло не так</div>;
+  }
+  const { data: movieInfo } = movieApi.useGetMovieQuery(parseInt(id));
 
   return (
     <div className={s.wrapper}>
@@ -31,12 +35,14 @@ export const MoviePage = () => {
                   <div className={s.descriptTitle}>Год:</div>
                   <div className={s.descripText}>{movieInfo.year}</div>
                 </div>
-                <div className={s.otherInfo}>
-                  <div className={s.descriptTitle}>Бюджет:</div>
-                  <div
-                    className={s.descripText}
-                  >{`${movieInfo.budget.value}${movieInfo.budget.currency}`}</div>
-                </div>
+                {movieInfo.budget.value && (
+                  <div className={s.otherInfo}>
+                    <div className={s.descriptTitle}>Бюджет:</div>
+                    <div
+                      className={s.descripText}
+                    >{`${movieInfo.budget.value}${movieInfo.budget.currency}`}</div>
+                  </div>
+                )}
               </div>
             </div>
             <div className={s.rightPart}>
@@ -44,6 +50,9 @@ export const MoviePage = () => {
               <div className={s.description}>
                 <div className={s.descriptTitle}>Описание:</div>
                 <div className={s.descripText}>{movieInfo.description}</div>
+              </div>
+              <div className={s.buttons}>
+                <FavouriteButton movieId={parseInt(id)} />
               </div>
             </div>
           </div>
